requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
fastapi==0.110.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-cli==0.0.21
langchain-community==0.0.31
langchain-core==0.1.40
langchain-text-splitters==0.0.1
langchain_openai
langserve==0.0.51
langsmith==0.1.40
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==69.0.3
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sse-starlette==1.8.2
starlette==0.37.2
tenacity==8.2.3
tomlkit==0.12.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.23.2
wheel==0.43.0
yarl==1.9.4
python-dotenv
openai==1.16.2